+++ src/converters/piral-vue-3/convert.ts
@@ -16,14 +16,15 @@ export interface Vue3Converter {
 export function createVue3Converter(...params: Parameters<typeof createConverter>) {
   const convert = createConverter(...params);
   const Extension = convert.Extension;
+  const defineMiddleware = convert.defineMiddleware;
   const from: Vue3Converter = (root, captured) => ({
     type: 'html',
     component: convert(root, captured),
   });
 
-  return { from, Extension };
+  return { from, Extension, defineMiddleware };
 }
 
-const { from: fromVue3, Extension: Vue3Extension } = createVue3Converter();
+const { from: fromVue3, Extension: Vue3Extension, defineMiddleware: defineVue3Middleware } = createVue3Converter();
 
-export { fromVue3, Vue3Extension };
+export { fromVue3, Vue3Extension, defineVue3Middleware };