+++ src/converters/piral-vue-3/src/converter.ts
@@ -2,6 +2,7 @@ import type { ForeignComponent, BaseComponentProps } from 'piral-core';
 import { Component, App } from 'vue';
 import { createExtension } from './extension';
 import { mountVue } from './mount';
+import { Vue3MiddlewareHandler } from './types';
 
 export interface Vue3ConverterOptions {
   /**
@@ -23,6 +24,7 @@ interface Vue3State {
 export function createConverter(config: Vue3ConverterOptions = {}) {
   const { rootName = 'piral-slot', selector = 'extension-component' } = config;
   const Extension = createExtension(rootName);
+  let middleware: Vue3MiddlewareHandler = () => {};
   const convert = <TProps extends BaseComponentProps>(
     root: Component<TProps>,
     captured?: Record<string, any>,
@@ -30,6 +32,7 @@ export function createConverter(config: Vue3ConverterOptions = {}) {
     mount(parent, data, ctx, locals: Vue3State) {
       const el = parent.appendChild(document.createElement(rootName));
       const app = mountVue(root, data, ctx, captured);
+      middleware(app);
       app.component(selector, createExtension(rootName));
       app.mount(el);
       !app._props && (app._props = {});
@@ -47,5 +50,8 @@ export function createConverter(config: Vue3ConverterOptions = {}) {
     },
   });
   convert.Extension = Extension;
+  convert.defineMiddleware = (setup: Vue3MiddlewareHandler) => {
+    middleware = setup;
+  };
   return convert;
 }